feat(tools): replace rename tool with unified session tool
Single 'session' tool that gets current title (no args) or renames
(with title arg). Rename is skipped if the user already set a custom
title, preventing the model from overwriting intentional names.
