+++ packages/opencode/src/session/system.ts
@@ -60,8 +60,8 @@ export const layer = Layer.effect(
             `</env>`,
             ``,
             `# Session Naming`,
-            `Once you understand the user's task, call the rename tool early with a short descriptive title (3-7 words).`,
-            `If the task changes significantly, rename again to keep the title accurate.`,
+            `Once you understand the user's task, call the session tool early with a short descriptive title (3-7 words).`,
+            `If the task changes significantly, call session again to update the title.`,
           ].join("\n"),
         ]
       }),
