+++ packages/opencode/src/tool/registry.ts
@@ -9,7 +9,7 @@ import { GrepTool } from "./grep"
 import { ReadTool } from "./read"
 import { TaskTool } from "./task"
 import { TodoWriteTool } from "./todo"
-import { RenameTool } from "./rename"
+import { SessionTool } from "./session"
 import { WebFetchTool } from "./webfetch"
 import { WriteTool } from "./write"
 import { InvalidTool } from "./invalid"
@@ -116,7 +116,7 @@ export const layer: Layer.Layer<
     const read = yield* ReadTool
     const question = yield* QuestionTool
     const todo = yield* TodoWriteTool
-    const rename = yield* RenameTool
+    const session = yield* SessionTool
     const cron = yield* Effect.promise(() => import("./cron")).pipe(Effect.map((mod) => mod.CronTool), Effect.flatten)
     const lsptool = yield* LspTool
     const plan = yield* PlanExitTool
@@ -224,7 +224,7 @@ export const layer: Layer.Layer<
           task: Tool.init(task),
           fetch: Tool.init(webfetch),
           todo: Tool.init(todo),
-          rename: Tool.init(rename),
+          session: Tool.init(session),
           cron: Tool.init(cron),
           search: Tool.init(websearch),
           code: Tool.init(codesearch),
@@ -252,7 +252,7 @@ export const layer: Layer.Layer<
             tool.task,
             tool.fetch,
             tool.todo,
-            tool.rename,
+            tool.session,
             tool.cron,
             tool.search,
             ...(Flag.OPENCODE_EXPERIMENTAL_SCOUT ? [tool.code, tool.repo_clone, tool.repo_overview] : []),
